refactor: clean up html logic
