--- ts/display/html/hexagon.html.ts
@@ -1,5 +1,0 @@
-export const generateHexagonHTML = (hexagonNumber: number) =>
-  `<div class="hexagon__outer">
-	  <div class="hexagon__inner">${hexagonNumber}</div>
-	</div>
-  `;
